Teach hbsd-update-build about cleanroom builds
Add an option to make hbsd-update-build checkout /usr/src from git
and build it before distributing to the chroot dir. This makes
hbsd-update-build work on systems that come without /usr/src
or without corresponding objects to distribute.

It is off by default (preserving current behavior) and can be enabled with
WANT_CHROOT_BUILD=1