HBSD: Handle Cross-DSO CFI runtime intercepting libc calls to dlopen
This commit provides an alternative way to call dlopen and dlclose in
the RTLD. This alternative isn't really that great of one: an attacker
could simply call our alternative rather than the official API,
effectively bypassing Cross-DSO CFI.

However, having this alternative allows us to make tangible progress
since simple applications now mostly work.

While here, fix a bug with llvm's Cross-DSO CFI implementation's
fallback implementation of mremap. FreeBSD lacks mremap, so the
Cross-DSO CFI runtime needs to copy the stored DSO metadata when
installing a new shadow map. The problem here, though, is that the size
of the shadow map is HUGE (larger than the physical memory of the VM I'm
doing development on). Using memmove/memcpy causes the kernel to try to
page in the entire memory allocation (which was mostly unused),
eventually leading the application to crash due to memory constraints.

So, as a follow-up work, we will likely need to provide an in-kernel
implementation of mremap.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>