vmstat: fix overflow of interrupt name buffer
sysctl() provides a count of number of bytes in the buffer.  That is the
actual buffer length.  Whereas looking for an interrupt entry with an
empty name could terminate too early, or overflow the end of the buffer.
The overflow will occur if the table of interrupt names is full.

Reviewed by:	kib
MFC after:	1 week
Differential Revision:	https://reviews.freebsd.org/D36628

(cherry picked from commit e5c30ac93055e708e26e075937263608b3eeb17d)
Signed-off-by: Shawn Webb <shawn.webb@hardenedbsd.org>