HBSD: Initial bulk import of Offensive-OpenSSL @ 365db85216493dfea4f0f5d05d4cd213d85a9c77
This relaxes a lot of sanity checks in OpenSSL's libcrypto. Effectively,
this causes libcrypto to accept all sorts of certificates, both valid
and invalid alike.

Eventually, we'll be able to generate all sorts of interesting
cryptographic key material. For example, an X509 certificate that claims
to have an EC pubkey, but in reality has an RSA pubkey.

Signed-off-by:	Shawn Webb <shawn.webb@ioactive.com>
Sponsored-by:	IOActive, Inc