HBSD: New data structures and core API for libhbsdcontrol
hbsdcontrol(8) uses libhbsdcontrol to handle performing the actual
toggling of security features on a per-file basis. The underlying design
of libhbsdcontrol and hbsdcontrol leave a bit to be desired. It's
currently difficult to write bindings for non-C languages.

This commit begins the rewrite of libhbsdcontrol. Attention is paid to
the data structures that govern the toggling of features.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
MFC-to:		14-STABLE
MFC-to:		13-STABLE