HBSD: Add more APIs and ABIs for libhbsdcontrol
A given file's feature state might be one of:

1. Enabled
2. Disabled
3. System default
4. Unknown
5. Invalid

The vast majority of times, getting the state should return system
default. A state is unknown when a system error occurs (perhaps
filesystem extended attributes aren't available for the underlying
filesystem.) A state of error occurs when an invalid combination of
settings occurs. For example, attempting to set both the enabled and
disabled filesystem extended attributes for a given file.

This also starts the ASLR feature toggle. Provide the implementation for
getting the feature state of a given file.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>