HBSD: Pass in a zero-entry environment to _execve in posix_spawn*
Some users of the posix_spawn(3) API pass in a NULL environment pointer
(envp). This causes the core of the posix_spawn API to inherit the
program's main environment (the environ global variable).

This might not be desirable in untrusted environments. Instead, we are
going to pass in an empty environment. Thus, users/consumers of
posix_spawn(3) must make the explicit decision for the executable image
to inherit the environment.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Issue:		#106
MFC-to:		14-STABLE
MFC-with:	cd93be7afbcfd134b45b52961fc9c6907984c85f